added method _savingLocation
